Split Marc21 and non-Marc21 SRU readers, add CERL Thesaurus
