+++ edpop_explorer/readers/hpb.py
@@ -1,7 +1,7 @@
-from edpop_explorer.srureader import SRUReader, SRURecord
+from edpop_explorer.srumarc21reader import SRUMarc21Reader, Marc21Record
 
 
-class HPBReader(SRUReader):
+class HPBReader(SRUMarc21Reader):
     sru_url = 'http://sru.k10plus.de/hpb'
     sru_version = '1.1'
     HPB_LINK = 'http://hpb.cerl.org/record/{}'
@@ -9,7 +9,7 @@ class HPBReader(SRUReader):
     def transform_query(self, query: str) -> str:
         return query
 
-    def get_link(self, record: SRURecord):
+    def get_link(self, record: Marc21Record):
         # The record id can be found in field 035 in subfield a starting
         # with (CERL), like this: (CERL)HU-SzSEK.01.bibJAT603188.
         # The URI can then be created using HPB_URI.
