+++ edpop_explorer/readers/vd.py
@@ -1,7 +1,7 @@
-from edpop_explorer.srureader import SRUReader, SRURecord
+from edpop_explorer.srumarc21reader import SRUMarc21Reader, Marc21Record
 
 
-class VD16Reader(SRUReader):
+class VD16Reader(SRUMarc21Reader):
     sru_url = 'http://bvbr.bib-bvb.de:5661/bvb01sru'
     sru_version = '1.1'
     VD16_LINK = 'http://gateway-bayern.de/{}'  # Spaces should be replaced by +
@@ -11,7 +11,7 @@ class VD16Reader(SRUReader):
         # queried
         return 'VD16 and ({})'.format(query)
 
-    def get_link(self, record: SRURecord):
+    def get_link(self, record: Marc21Record):
         try:
             identifier = record.get_first_field('024').subfields['a']
         except KeyError:
@@ -19,7 +19,7 @@ class VD16Reader(SRUReader):
         return self.VD16_LINK.format(identifier).replace(' ', '+')
 
 
-class VD17Reader(SRUReader):
+class VD17Reader(SRUMarc21Reader):
     sru_url = 'http://sru.k10plus.de/vd17'
     sru_version = '1.1'
     VD17_LINK = \
@@ -28,7 +28,7 @@ class VD17Reader(SRUReader):
     def transform_query(self, query: str) -> str:
         return query
 
-    def get_link(self, record: SRURecord):
+    def get_link(self, record: Marc21Record):
         try:
             identifier = record.get_first_field('024').subfields['a']
         except KeyError:
@@ -36,7 +36,7 @@ class VD17Reader(SRUReader):
         return self.VD17_LINK.format(identifier)
 
 
-class VD18Reader(SRUReader):
+class VD18Reader(SRUMarc21Reader):
     sru_url = 'http://sru.k10plus.de/vd18'
     sru_version = '1.1'
     VD18_LINK = 'https://kxp.k10plus.de/DB=1.65/SET=1/TTL=1/CMD?ACT=SRCHA&' \
@@ -46,7 +46,7 @@ class VD18Reader(SRUReader):
     def transform_query(self, query: str) -> str:
         return query
 
-    def get_link(self, record: SRURecord):
+    def get_link(self, record: Marc21Record):
         # The record id is in field 024 for which subfield 2 is vd18. There
         # may be more than one occurance of field 024.
         fields024 = record.get_fields('024')
