+++ edpop_explorer/srureader.py
@@ -1,92 +1,8 @@
 import sruthi
-import csv
-from pathlib import Path
-from typing import Dict, List, Optional
-from dataclasses import dataclass, field as dataclass_field
+from typing import List
+import json
 
-from edpop_explorer.apireader import APIReader
-
-
-READABLE_FIELDS_FILE = Path(__file__).parent / 'M21_fields.csv'
-translation_dictionary = {}
-with open(READABLE_FIELDS_FILE) as dictionary_file:
-    reader = csv.DictReader(dictionary_file)
-    for row in reader:
-        translation_dictionary[row['Tag number']] = \
-            row[' Tag description'].strip()
-
-
-@dataclass
-class SRURecordField:
-    fieldnumber: str
-    indicator1: str
-    indicator2: str
-    subfields: Dict[str, str] = dataclass_field(default_factory=list)
-    description: Optional[str] = None
-
-    def __repr__(self):
-        '''
-        Return the usual marc21 representation
-        '''
-        sf = []
-        ind1 = self.indicator1 if self.indicator1.rstrip() != '' else '#'
-        ind2 = self.indicator1 if self.indicator2.rstrip() != '' else '#'
-        description = ' ({})'.format(self.description) \
-            if self.description else ''
-        for subfield in self.subfields:
-            sf.append('$${} {}'.format(subfield, self.subfields[subfield]))
-        return '{}{}: {} {} {}'.format(
-            self.fieldnumber,
-            description,
-            ind1,
-            ind2,
-            '  '.join(sf)
-        )
-
-
-@dataclass
-class SRURecord:
-    # We use a list for the fields and not a dictionary because they may
-    # appear more than once
-    fields: List[SRURecordField] = dataclass_field(default_factory=list)
-    link: Optional[str] = None
-
-    def get_first_field(self, fieldnumber: str) -> SRURecordField:
-        '''Return the first occurance of a field with a given field number.
-        May be useful for fields that appear only once, such as 245.
-        Return None if field is not found.'''
-        for field in self.fields:
-            if field.fieldnumber == fieldnumber:
-                return field
-        return None
-
-    def get_fields(self, fieldnumber: str) -> List[SRURecordField]:
-        '''Return a list of fields with a given field number. May return an
-        empty list if field does not occur.'''
-        returned_fields = []
-        for field in self.fields:
-            if field.fieldnumber == fieldnumber:
-                returned_fields.append(field)
-        return returned_fields
-
-    def get_title(self):
-        field_245 = self.get_first_field('245')
-        if field_245:
-            return field_245.subfields.get('a', '(unknown title)')
-        else:
-            return '(unknown title)'
-
-    def show_record(self) -> str:
-        field_strings = []
-        if self.link:
-            field_strings.append(self.link)
-        for field in self.fields:
-            field_strings.append(str(field))
-        return '\n'.join(field_strings)
-
-    def __repr__(self):
-        return self.get_title()
-
+from edpop_explorer.apireader import APIReader, APIRecord
 
 RECORDS_PER_PAGE = 10
 
@@ -95,16 +11,19 @@ class SRUReader(APIReader):
     sru_url: str = None
     sru_version: str = None
     query: str = None
-    records: List[SRURecord]
+    records: List[APIRecord]  # Move to superclass?
     fetching_exhausted: bool = False
 
     def transform_query(self, query: str) -> str:
         raise NotImplementedError('Should be implemented by subclass')
 
-    def get_link(self, record: SRURecord) -> str:
+    def get_link(self, record: APIRecord) -> str:
         raise NotImplementedError('Should be implemented by subclass')
 
-    def _perform_query(self, query: str, start_record: int) -> List[SRURecord]:
+    def _convert_record(self, sruthirecord: dict) -> APIRecord:
+        raise NotImplementedError('Should be implemented by subclass')
+
+    def _perform_query(self, query: str, start_record: int) -> List[dict]:
         try:
             response = sruthi.searchretrieve(
                 self.sru_url,
@@ -118,34 +37,12 @@ class SRUReader(APIReader):
         ):
             raise
 
-        records: List[SRURecord] = []
         self.number_of_results = response.count
 
+        records: List[APIRecord] = []
         for sruthirecord in response[0:RECORDS_PER_PAGE]:
-            record = SRURecord()
-            for sruthifield in sruthirecord['datafield']:
-                fieldnumber = sruthifield['tag']
-                field = SRURecordField(
-                    fieldnumber=fieldnumber,
-                    indicator1=sruthifield['ind1'],
-                    indicator2=sruthifield['ind2'],
-                    subfields={}
-                )
-                if fieldnumber in translation_dictionary:
-                    field.description = translation_dictionary[fieldnumber]
-                # If there are multiple subfields, sruthi puts it directly in
-                # a dict, otherwise it uses a list of dicts
-                if type(sruthifield['subfield']) == dict:
-                    sruthisubfields = [sruthifield['subfield']]
-                else:
-                    sruthisubfields = sruthifield['subfield']
-                assert type(sruthisubfields) == list
-                for sruthisubfield in sruthisubfields:
-                    field.subfields[sruthisubfield['code']] = \
-                        sruthisubfield['text']
-                record.fields.append(field)
-            record.link = self.get_link(record)
-            records.append(record)
+            records.append(self._convert_record(sruthirecord))
+
         return records
 
     def fetch(self, query: str) -> None:
